Add ApplicationGateway, ApplicationManager's functions now use domain entities

+++ src/application/application_gateway/dto/shop_dto.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <string>
 #include <chrono>
+#include <cstdint>
 #include <string>
 
 namespace application {
@@ -12,14 +12,14 @@ struct AddEmployeeInputDto {
     std::string patronymic;
     std::string birth_date;
     std::string employment_date;
-    int employee_number;
-    int department_id;
-    int staff_position_id;
-    int work_schedule_id;
+    uint64_t employee_number;
+    uint64_t department_id;
+    uint64_t staff_position_id;
+    uint64_t work_schedule_id;
 };
 
 struct EmployeeIdInputDto {
-    int employee_id;
+    uint64_t employee_id;
 };
 
 struct EmployeeOutputDto {
@@ -28,7 +28,7 @@ struct EmployeeOutputDto {
     std::optional<std::string> patronymic;
     std::chrono::sys_days birth_date;
     std::chrono::sys_days employment_date;
-    unsigned int employee_number;
+    uint64_t employee_number;
 };
 
-} // namespace application
+}  // namespace application
